Initial commit
This is the initial commit for the read- and write- twitter interface.

For a given user (set via a YAML config file), it reads the DMs. If the DMs are from an allowed sender (also from that YAML config file), it saves that DM to a database. This is set via crontab.

At regular intervals (also via crontab) the oldest unread DM gets posted to the user's Twitter account. This allows there to be a steady, regular output of that Twitter account. This is being used for the Twitter feed for the Purdue Perl Mongers, @purduepm